+++ packages/author-profile/author-profile-content.web.js
@@ -16,9 +16,8 @@ const AuthorProfile = props =>
   <View>
     <AuthorProfileHeader {...props} />
     {props.articles.list.map((item, key) => {
-      const separatorComponent = key > 0
-        ? <AuthorProfileItemSeparator />
-        : null;
+      const separatorComponent =
+        key > 0 ? <AuthorProfileItemSeparator /> : null;
 
       return (
         <View key={item.id} style={styles.container}>