+++ packages/caption/CHANGELOG.md
@@ -3,6 +3,12 @@
 All notable changes to this project will be documented in this file.
 See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+<a name="0.2.3"></a>
+## [0.2.3](https://github.com/newsuk/times-components/compare/@times-components/caption@0.2.2...@times-components/caption@0.2.3) (2017-08-22)
+
+
+
+
 <a name="0.2.2"></a>
 ## [0.2.2](https://github.com/newsuk/times-components/compare/@times-components/caption@0.2.1...@times-components/caption@0.2.2) (2017-08-17)